requirements.txt
requests
zyte-api>=0.7.0
beautifulsoup4
lxml>=4.6.0
dataclasses-json
python-dotenv
